numbers file resolution is at hand
It's still borked in this commit, but it's now
"working" sensibly (just broken).

The issue was almost surely a conflict with an internal
`numbers` module as part of python. I don't know the python
lib very well yet, which is part of why I'm writing these
little python programs after all.